requirements-release.txt
aiohttp==3.5.4
galaxy.plugin.api==0.42
